+++ NutriApp/Views/AnthropometricEvaluation/AnthropometricEvaluationPage.xaml
@@ -14,6 +14,7 @@
     xmlns:anthropometricEvaluation="clr-namespace:NutriApp.Views.AnthropometricEvaluation"
     xmlns:inputLayout="clr-namespace:MauiLib1.Components.InputLayout;assembly=BaseMauiLib"
     xmlns:contentPageCustomer="clr-namespace:NutriApp.Components.ContentPageCustomer"
+    xmlns:resultView="clr-namespace:NutriApp.Views.AnthropometricEvaluation.View.ResultView"
     TypeTitle="None"
     TextTitle="Avaliação Antropométrica"
     x:Class="NutriApp.Views.AnthropometricEvaluation.AnthropometricEvaluationPage">
@@ -126,7 +127,7 @@
 
                 </toolkit:Expander>
 
-                <toolkit:Expander>
+                <toolkit:Expander Margin="0,0,0,10">
 
                     <toolkit:Expander.Header>
                         <Frame
@@ -144,26 +145,31 @@
                     </toolkit:Expander.Header>
 
                     <bodyComposition:BodyCompositionView
+                        x:Name="BodyCompositionViewTeste"
+                        GenderType="{Binding GenderType.Nome}"
+                        AnthropometricEvaluationTypeString="{Binding AnthropometricEvaluationType}"
+                        AnthropometricEvaluationModel="{Binding BodyCompositionPageModel, Mode=TwoWay}"
                         Margin="0,0,0,0" />
+                    
                 </toolkit:Expander>
 
                 <BoxView
+                    Margin="0,10,0,10"
                     Color="{StaticResource DarkBlue}"
                     HeightRequest="2"
                     IsVisible="{Binding CanDisplayResult}"
                     HorizontalOptions="Fill" />
 
-                <!-- <resultView:AnthropometricEvaluationResultView -->
-                <!--     Margin="0,20,0,0" -->
-                <!--     x:Name="AnthropometricEvaluationResultView" -->
-                <!--     IsVisible="{Binding CanDisplayResult}" -->
-                <!--     BoneDiameterModel="{Binding BoneDiameterModel}" -->
-                <!--     AnthropometricEvaluationResultModel="{Binding AnthropometricEvaluationResultModel}" /> -->
+                <resultView:AnthropometricEvaluationResultView
+                    x:Name="AnthropometricEvaluationResultView"
+                    IsVisible="{Binding CanDisplayResult}"
+                    BoneDiameterModel="{Binding BoneDiameterModel}"
+                    AnthropometricEvaluationResultModel="{Binding AnthropometricEvaluationResultModel}" />
 
             </StackLayout>
 
             <button:ButtonView
-                Margin="0,10,0,0"
+                Margin="0,0,0,0"
                 VerticalOptions="CenterAndExpand"
                 TextColor="White"
                 BackgroundColor="{StaticResource DarkBlue}"
