+++ NutriApp/Views/AnthropometricEvaluation/AnthropometricEvaluationPage.xaml.cs
@@ -7,6 +7,11 @@ public partial class AnthropometricEvaluationPage : BaseContentPage
     public AnthropometricEvaluationPage()
     {
         InitializeComponent();
-        BindingContext = new AnthropometricEvaluationViewModel();
+        BindingContext = new AnthropometricEvaluationViewModel(this);
+    }
+
+    public bool ValidateViewData()
+    {
+        return BodyCompositionViewTeste.ValidateViewData();
     }
 }
